backend/requirements.txt
fastapi
uvicorn
pillow
reportlab
python-multipart